64 bit fixes for the .gui framework.
git-svn-id: svn://0x263a.de/tekuum_src@374 3ae3452d-76a7-4814-9b33-b51e13292e42

+++ ui/Window.h
@@ -255,8 +255,9 @@ public:
 	idStr* GetStrPtrByName( const char* _name );
 	
 	virtual idWinVar* GetWinVarByName( const char* _name, bool winLookup = false, drawWin_t** owner = NULL );
-	
-	int  GetWinVarOffset( idWinVar* wv, drawWin_t* dw );
+	// DG: the return value is a pointer, so use intptr_t
+	intptr_t GetWinVarOffset( idWinVar* wv, drawWin_t* dw );
+	// DG end
 	float GetMaxCharHeight();
 	float GetMaxCharWidth();
 	void SetFont();
@@ -332,7 +333,9 @@ public:
 	bool RunScript( int n );
 	bool RunScriptList( idGuiScriptList* src );
 	void SetRegs( const char* key, const char* val );
-	int ParseExpression( idParser* src, idWinVar* var = NULL, int component = 0 );
+	// DG: component and the return value are really pointers, so use intptr_t
+	intptr_t ParseExpression( idParser* src, idWinVar* var = NULL, intptr_t component = 0 );
+	// DG end
 	int ExpressionConstant( float f );
 	idRegisterList* RegList()
 	{
@@ -393,10 +396,12 @@ protected:
 	
 	int ExpressionTemporary();
 	wexpOp_t* ExpressionOp();
-	int EmitOp( int a, int b, wexpOpType_t opType, wexpOp_t** opp = NULL );
-	int ParseEmitOp( idParser* src, int a, wexpOpType_t opType, int priority, wexpOp_t** opp = NULL );
-	int ParseTerm( idParser* src, idWinVar* var = NULL, int component = 0 );
-	int ParseExpressionPriority( idParser* src, int priority, idWinVar* var = NULL, int component = 0 );
+	// DG: a, b, component and the return values are really pointers, so use intptr_t
+	intptr_t EmitOp( intptr_t a, intptr_t b, wexpOpType_t opType, wexpOp_t** opp = NULL );
+	intptr_t ParseEmitOp( idParser* src, intptr_t a, wexpOpType_t opType, int priority, wexpOp_t** opp = NULL );
+	intptr_t ParseTerm( idParser* src, idWinVar* var = NULL, intptr_t component = 0 );
+	intptr_t ParseExpressionPriority( idParser* src, int priority, idWinVar* var = NULL, intptr_t component = 0 );
+	// DG end
 	void EvaluateRegisters( float* registers );
 	void SaveExpressionParseState();
 	void RestoreExpressionParseState();
